reinstate file watcher for plots files
