+++ extension/src/plots/data/index.ts
@@ -5,48 +5,33 @@ import { PlotsOutput } from '../../plots/webview/contract'
 import { sameContents } from '../../util/array'
 
 export class PlotsData extends BaseData<PlotsOutput> {
-  private revisions?: string[]
-
   constructor(
     dvcRoot: string,
     internalCommands: InternalCommands,
     updatesPaused: EventEmitter<boolean>
   ) {
-    super(dvcRoot, internalCommands, updatesPaused)
+    super(
+      dvcRoot,
+      internalCommands,
+      AvailableCommands.PLOTS_DIFF,
+      updatesPaused
+    )
+  }
 
-    this.initialize()
+  public collectFiles(data: PlotsOutput) {
+    return Object.keys(data)
   }
 
   public clearRevisions() {
-    this.revisions = undefined
+    this.args = undefined
   }
 
-  public setRevisions(...revisions: string[]) {
-    if (this.revisions && sameContents(revisions, this.revisions)) {
+  public setRevisions(...args: string[]) {
+    if (this.args && sameContents(args, this.args)) {
       return
     }
 
-    this.revisions = revisions
+    this.args = args
     this.managedUpdate()
-  }
-
-  public async update(): Promise<void> {
-    const data = await this.internalCommands.executeCommand<PlotsOutput>(
-      AvailableCommands.PLOTS_DIFF,
-      this.dvcRoot,
-      ...(this.revisions || [])
-    )
-
-    return this.notifyChanged(data)
-  }
-
-  private initialize() {
-    const waitForInitialData = this.dispose.track(
-      this.onDidUpdate(() => {
-        this.dispose.untrack(waitForInitialData)
-        waitForInitialData.dispose()
-        this.deferred.resolve()
-      })
-    )
   }
 }
